+++ .github/scripts/cargo_build.sh
@@ -72,7 +72,7 @@ fi
 find . -type d -name cosmian-kms -exec rm -rf \{\} \; -print || true
 rm -f /tmp/*.json
 
-export RUST_LOG="cosmian_kms_cli=error,cosmian_kms_server=debug"
+export RUST_LOG="cosmian_kms_cli=warn,cosmian_kms_server=debug"
 # shellcheck disable=SC2086
 cargo test --target $TARGET $RELEASE $FEATURES --workspace -- --nocapture $SKIP_SERVICES_TESTS